Update README.md
I think this is a great idea and I'm sure you will find a few things. Please work on your writing for clarity and don't assume the reader understands what you mean---i.e., assume your audience doesn't play golf.  Also, be careful with your causal language. Just because a given metric is associated with performance, doesn't mean it causes high performance. We may need to narrow your question to specific parts of the game once we have a better understanding of the data.

87/100